+++ src/GRA/SiteSettingKey.cs
@@ -58,6 +58,9 @@
             public static readonly string DefaultDailyPersonalGoal
                 = "Users.DefaultDailyPersonalGoal";
 
+            public static readonly string MaximumActivityPermitted
+                = "Users.MaximumActivityPermitted";
+
             public static readonly string SurveyUrl = "Users.SurveyUrl";
             public static readonly string FirstTimeSurveyUrl = "Users.FirstTimeSurveyUrl";
             public bool Equals(Users other) { return true; }